Refactor resiliency as an extension of client
